% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcases-simulate.R
\name{simcases_simulate}
\alias{simcases_simulate}
\title{Simulate Datasets for multiple cases/scenarios}
\usage{
simcases_simulate(code, constants = list(nlist()),
  parameters = list(nlist()), monitor = list(".*"), cases = list(""),
  nsims, parallel = 1, path = NULL, exists = FALSE, silent = FALSE)
}
\arguments{
\item{code}{A list of strings of JAGS or R code to generate the data. The code must not be in a data or model block.}

\item{constants}{A list of nlists objects specifying the values of nodes in code. The values are included in the output data.}

\item{parameters}{A list of nlist objects specifying the values of nodes in code. The values are not included in the output data.}

\item{monitor}{A list of character vectors (or regular expression if a string) specifying the names of the stochastic nodes in code to include in the data. By default all stochastic nodes are included.}

\item{cases}{A list of strings specifying the cases for which data shall be simulated.}

\item{nsims}{A vector of integers specifying the number of data sets to simulate for each case. By default 100 data sets are simulated for each case.}

\item{parallel}{An integer specifying the number of CPU cores to use for generating the datasets in parallel. Defaul is 1 (not parallel).}

\item{path}{A string specifying the path to the directory to save the data sets in. By default \code{path = NULL } the data sets are not saved but are returned as an nlists object.}

\item{exists}{A flag specifying whether the directory should already exist. If \code{exists = NA} it doesn't matter. If the directory already exists it is overwritten if \code{exists = TRUE} or \code{exists = NA} otherwise an error is thrown.}

\item{silent}{A flag specifying whether to suppress warnings.}
}
\value{
A flag.
}
\description{
Simulate Datasets for multiple cases/scenarios
}
\examples{
simcases_simulate(code = list(code = "a ~ dnorm(0, 1/sigma^2)"),
                  constants = list(const1 = nlist(sigma=1), 
                                 const2 = nlist(sigma=5)),
                  cases = list(data1 = "code const1",
                             data2 = "code const2"))
}
