% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse-bayesian.R
\name{smc_analyse_bayesian}
\alias{smc_analyse_bayesian}
\title{Analyse Datasets for multiple cases/scenarios using Bayesian methods}
\usage{
smc_analyse_bayesian(models, cases, path = ".",
  environment = parent.frame(), fun = function(x) read.table(text =
  gsub(";|,| |:|\\t|\\\\||&|~", "\\t", readLines(textConnection(x))),
  header = TRUE), ...)
}
\arguments{
\item{models}{An object that becomes a data frame when \code{fun} is applied. The first row of the data frame is a header and each subsequent row defines a case using strings. The strings must refer to objects defined within \code{environment}.}

\item{cases}{An object that becomes a data frame when \code{fun} is applied. The first row of the data frame is a header and each subsequent row defines a case using strings. The strings must refer to objects defined within \code{environment}.}

\item{path}{A string specifying the path to the directory.}

\item{environment}{The environment in which the objects described in \code{cases} were defined.}

\item{fun}{A function to convert \code{models} and \code{cases} to data frames.}

\item{...}{Other arguments of \code{sims::sims_simulate}}
}
\description{
Analyse Datasets for multiple cases/scenarios using Bayesian methods. Results are saved to disk.
}
