% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_set_case.R
\name{smc_set_case}
\alias{smc_set_case}
\title{Specify a case/scenario}
\usage{
smc_set_case(fun, ...)
}
\arguments{
\item{fun}{A function for a case.}

\item{...}{Arguments to pass to \code{function} for a case}
}
\description{
Specify a case/scenario
}
\examples{
smc_set_case(fun=sims::sims_simulate, 
             code = "a ~ dnorm(0, 1/sigma^2)",
             constants = nlist(sigma=1))
}
