% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcases-simulate.R
\name{smc_simulate}
\alias{smc_simulate}
\title{Simulate data from various models.}
\usage{
smc_simulate(
  models = models,
  code = "",
  constants = nlist::nlist(),
  parameters = nlist::nlist(),
  monitor = ".*",
  stochastic = TRUE,
  latent = FALSE,
  nsims = 1,
  path = ".",
  exists = FALSE,
  rdists = sims_rdists(),
  ask = getOption("sims.ask", TRUE),
  silent = FALSE,
  environment = parent.frame(),
  fun = function(x) read.table(text = gsub(";|,| |:|\\t|\\\\||&|~", "\\t",
    readLines(textConnection(x))), header = TRUE),
  ...
)
}
\arguments{
\item{models}{A data frame or an object that becomes a data frame when \code{fun} is applied. 
The data frame contains character strings that refer to objects defined within \code{environment}. 
Each row defines a model. 
The header may contain: code, constants, parameters, monitor, stochastic, latent and/or nsims}

\item{code}{A string of the JAGS or R code to generate the data.
The JAGS code must not be in a data or model block.}

\item{constants}{An nlist object (or list that can be coerced to nlist)
specifying the values of nodes in code.
The values are included in the output dataset.}

\item{parameters}{An nlist object (or list that can be coerced to nlist)
specifying the values of nodes in code.
The values are not included in the output dataset.}

\item{monitor}{A character vector (or regular expression if a string)
specifying the names of the nodes in code to include in the dataset.
By default all nodes are included.}

\item{stochastic}{A logical scalar specifying whether to monitor
deterministic and stochastic (NA), only deterministic (FALSE)
or only stochastic nodes (TRUE).}

\item{latent}{A logical scalar specifying whether to monitor
observed and latent (NA), only latent (TRUE)
or only observed nodes (FALSE).}

\item{nsims}{A whole number between 1 and 1,000,000 specifying
the number of data sets to simulate. By default 100 data sets are simulated.}

\item{path}{A string specifying the path to the directory to save the data sets in.}

\item{exists}{A flag specifying whether the `path` directory should already exist
(if `exists = NA` it doesn't matter).}

\item{rdists}{A character vector specifying the R functions to recognize as stochastic.
The seed should be specified using [base::set.seed()].}

\item{ask}{A flag specifying whether to ask before deleting sims compatible files.}

\item{silent}{A flag specifying whether to suppress warnings.}

\item{environment}{The environment in which the objects described in \code{cases} were defined.}

\item{fun}{A function to convert \code{cases} to a data frame.}

\item{...}{Unused.}
}
\description{
Simulate data from various models. Results are saved to disk.
}
\details{
The arguments \code{code}, \code{constants}, \code{parameters}, \code{monitor}, \code{stochastic}, \code{latent} and \code{nsims} will only be used if not specified within \code{models}.
}
\examples{

normal <- "a ~ dnorm(0, 1/sigma^2)"
sigma1 <- list(sigma=1)
sigma2 <- list(sigma=2)
smc_simulate("code parameters,
                   normal sigma1
                   normal sigma2",
                   path = tempdir(), 
                   exists = NA, 
                   ask = FALSE)
smc_simulate(tibble::tribble(
                   ~code, ~parameters,
                   "normal", "sigma1",
                   "normal", "sigma2"),
                   fun = identity,
                   path = tempdir(), 
                   exists = NA, 
                   ask = FALSE)
}
\seealso{
[sims::simulate()]
}
