% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgtemplate-package.R
\docType{package}
\name{simcases-package}
\alias{simcases}
\alias{simcases-package}
\title{simcases: Perform Simulation Studies Over Multiple Cases/Scenarios}
\description{
Perform Simulation Studies Over Multiple Cases/Scenarios.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/simcases}
  \item Report bugs at \url{https://github.com/poissonconsulting/simcases/issues}
}

}
\author{
\strong{Maintainer}: Audrey Beliveau \email{audrey.beliveau@uwaterloo.ca}

Authors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (0000-0002-7683-4592)
}

}
\keyword{internal}
