% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcases-simulate.R
\name{simcases_simulate}
\alias{simcases_simulate}
\title{Simulate Datasets for multiple cases/scenarios}
\usage{
simcases_simulate(cases)
}
\arguments{
\item{cases}{A list of cases obtained with smc_set_cases}
}
\description{
Simulate Datasets for multiple cases/scenarios
}
\examples{

code <- "a ~ dnorm(0, 1/sigma^2)"
case1 <- smc_set_case(fun=sims::sims_simulate, 
             code = code,
             constants = nlist(sigma=1))
case2 <- smc_set_case(fun=sims::sims_simulate, 
             code = code,
             constants = nlist(sigma=2))         
simcases_simulate(cases = list(case1, case2))
}
