% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcases-simulate.R
\name{simcases_simulate}
\alias{simcases_simulate}
\title{Simulate Datasets for multiple cases/scenarios}
\usage{
simcases_simulate(header, cases, environment = parent.frame())
}
\arguments{
\item{header}{A string to indicate which arguments are used to define cases}

\item{cases}{A string that defines cases in terms of the header. Each row represents a different case. Objects must be defined within the global environment.}

\item{environment}{The environment in which the objects were defined.}
}
\description{
Simulate Datasets for multiple cases/scenarios
}
\examples{

normal <- "a ~ dnorm(0, 1/sigma^2)"
sigma1 <- nlist(sigma=1)
sigma2 <- nlist(sigma=2)
simcases_simulate("code constants",
                       "normal sigma1
                        normal sigma2")
}
