% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcases-analyse.R
\name{smc_analyse}
\alias{smc_analyse}
\title{Analyse Datasets for multiple cases/scenarios using Bayesian methods}
\usage{
smc_analyse(
  models,
  cases,
  path = ".",
  environment = parent.frame(),
  fun = function(x) read.table(text = gsub(";|,| |:|\\t|\\\\||&|~", "\\t",
    readLines(textConnection(x))), header = TRUE),
  ...
)
}
\arguments{
\item{models}{An object that becomes a data frame when \code{fun} is applied. The first row of the data frame is a header and each subsequent row defines a case using strings. The strings must refer to objects defined within \code{environment}.}

\item{cases}{An object that becomes a data frame when \code{fun} is applied. The first row of the data frame is a header and each subsequent row defines a case using strings. The strings must refer to objects defined within \code{environment}.}

\item{path}{A string specifying the path to the directory.}

\item{environment}{The environment in which the objects described in \code{cases} were defined.}

\item{fun}{A function to convert \code{models} and \code{cases} to data frames.}

\item{...}{Other arguments from \code{simanalyse::sma_analyse}}
}
\description{
Analyse Datasets for multiple cases/scenarios using Bayesian methods. Results are saved to disk.
}
\examples{

normal <- "a ~ dnorm(0, 1/sigma^2)"
sigma1 <- list(sigma=1)
sigma2 <- list(sigma=2)
all <- ".*"
models_sims <- "code   constants
                normal sigma1
                normal sigma2"
smc_simulate(models = models_sims,
                  path = tempdir(), 
                  exists = NA, 
                  ask = FALSE,
                  nsims = 2)
models_analysis <- "code   monitor  
                    normal all
                    normal all"
cases <- "sims analyse
          1    1
          2    2"
smc_analyse(models = models_analysis,
                     cases = cases,
                     path = tempdir())
}
